requirements.txt
PyYAML==6.0
pandas==1.4.4
torch==1.11.0+cu113
numpy==1.23.3
transformers==4.22.1
seaborn==0.12.0
matplotlib==3.6.0
joblib==1.2.0
sklearn==0.0
scikit-learn==1.1.2
Flask==2.2.2